BREAKING CHANGE New ActorEx, PlayerEx, and NpcEx wrappers and query structures.
All wrapper types now appropriately implement Intractable and/or Locatable with some nice implemented stuff right on them now such as .interact() methods.

More enhancements to come from this, however this is the big breaking change, so I'll push this now and a second part next as to rip off the band-aid.